feat(LWD): rename drawer files to dialog in ModularDialog
