+++ .changeset/silver-hounds-guess.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+Rename drawer files to dialog in ModularDialog
