+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/analytics/modularDialog.types.ts
