+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/hooks/__tests__/useModularDialogNavigation.test.ts
@@ -1,5 +1,5 @@
 import { renderHook } from "tests/testSetup";
-import { useModularDrawerNavigation } from "../useModularDrawerNavigation";
+import { useModularDrawerNavigation } from "../useModularDialogNavigation";
 import { act } from "react-dom/test-utils";
 
 describe("useModularDrawerNavigation", () => {
