+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/hooks/useDetailedAccounts.tsx
@@ -11,8 +11,8 @@ import { counterValueCurrencySelector } from "~/renderer/reducers/settings";
 import { BaseRawDetailedAccount } from "@ledgerhq/live-common/modularDrawer/types/detailedAccount";
 import { useDetailedAccountsCore } from "@ledgerhq/live-common/modularDrawer/hooks/useDetailedAccountsCore";
 import { isTokenCurrency } from "@ledgerhq/live-common/currencies/helpers";
-import { useModularDrawerAnalytics } from "../analytics/useModularDrawerAnalytics";
-import { MODULAR_DRAWER_PAGE_NAME } from "../analytics/modularDrawer.types";
+import { useModularDrawerAnalytics } from "../analytics/useModularDialogAnalytics";
+import { MODULAR_DRAWER_PAGE_NAME } from "../analytics/modularDialog.types";
 import { useOpenAssetFlowDialog } from "./useOpenAssetFlow";
 import { ModularDrawerLocation } from "LLD/features/ModularDrawer";
 import { Account, AccountLike } from "@ledgerhq/types-live";
