+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/hooks/useModularDialogData.ts
