+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/hooks/useModularDialogNavigation.ts
