+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/screens/AccountSelector/components/AccountSelectorContent/index.tsx
@@ -2,8 +2,8 @@ import React, { useMemo } from "react";
 import { Account, AccountLike } from "@ledgerhq/types-live";
 import { AccountVirtualList } from "../AccountVirtualList";
 import { ListWrapper } from "../../../../components/ListWrapper";
-import { useModularDrawerAnalytics } from "../../../../analytics/useModularDrawerAnalytics";
-import { MODULAR_DRAWER_PAGE_NAME } from "../../../../analytics/modularDrawer.types";
+import { useModularDrawerAnalytics } from "../../../../analytics/useModularDialogAnalytics";
+import { MODULAR_DRAWER_PAGE_NAME } from "../../../../analytics/modularDialog.types";
 import { AccountTuple } from "@ledgerhq/live-common/utils/getAccountTuplesForCurrency";
 import { BaseRawDetailedAccount } from "@ledgerhq/live-common/modularDrawer/types/detailedAccount";
 import { formatCurrencyUnit } from "@ledgerhq/coin-framework/currencies/formatCurrencyUnit";
