+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/screens/AssetSelector/index.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
 import SearchInputContainer from "./components/SearchInputContainer";
 import { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/wallet-api/ModularDrawer/types";
-import { MODULAR_DRAWER_PAGE_NAME } from "../../analytics/modularDrawer.types";
-import TrackDrawerScreen from "../../analytics/TrackDrawerScreen";
+import { MODULAR_DRAWER_PAGE_NAME } from "../../analytics/modularDialog.types";
+import TrackDrawerScreen from "../../analytics/TrackDialogScreen";
 import { LoadingStatus } from "@ledgerhq/live-common/deposit/type";
 import { GenericError } from "../../components/GenericError";
 import { useSelector } from "react-redux";
