+++ apps/ledger-live-desktop/src/newArch/features/ModularDialog/screens/NetworkSelector/index.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { NetworkSelectorContent } from "./components/NetworkSelectorContent";
 import { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/wallet-api/ModularDrawer/types";
-import TrackDrawerScreen from "../../analytics/TrackDrawerScreen";
-import { MODULAR_DRAWER_PAGE_NAME } from "../../analytics/modularDrawer.types";
+import TrackDrawerScreen from "../../analytics/TrackDialogScreen";
+import { MODULAR_DRAWER_PAGE_NAME } from "../../analytics/modularDialog.types";
 
 export type NetworkSelectorProps = {
   networks?: CryptoOrTokenCurrency[];
